refactor dw queries, update file names and change model datatypes

+++ service/src/main/kotlin/fi/turku/evakaturku/dw/DwModels.kt
@@ -163,7 +163,7 @@ data class DwDailyUnitOccupancyConfirmed(
     val pvm: LocalDate,
     val toimintayksikköId: UUID,
     val toimintayksikkö: String,
-    val kasvattajienLkm: Int?,
+    val kasvattajienLkm: Double?,
     val sijoituksienLkm: Int?,
     val täyttöasteSumma: Double?,
     val täyttöasteProsentteina: Double?,
@@ -173,9 +173,9 @@ data class DwDailyUnitOccupancyRealized(
     val pvm: LocalDate,
     val toimintayksikköId: UUID,
     val toimintayksikkö: String,
-    val kasvattajienLkm: Int?,
+    val kasvattajienLkm: Double?,
     val sijoituksienLkm: Int?,
-    val käyttöasteSumma: Int?,
+    val käyttöasteSumma: Double?,
     val käyttöasteProsentteina: Double?,
 )
